--- a/cpp/src/arrow/flight/flight-benchmark.cc
+++ b/cpp/src/arrow/flight/flight-benchmark.cc
@@ -93,8 +93,7 @@ Status RunPerformanceTest(const std::string& hostname, const int port) {
   RETURN_NOT_OK(plan->GetSchema(&schema));
 
   PerformanceStats stats;
-  auto ConsumeStream = [&stats, &schema, &hostname,
-                        &port](const FlightEndpoint& endpoint) {
+  auto ConsumeStream = [&stats, &hostname, &port](const FlightEndpoint& endpoint) {
     // TODO(wesm): Use location from endpoint, same host/port for now
     std::unique_ptr<FlightClient> client;
     RETURN_NOT_OK(FlightClient::Connect(hostname, port, &client));