--- a/cpp/src/parquet/column_scanner.h
+++ b/cpp/src/parquet/column_scanner.h
@@ -89,7 +89,7 @@ class PARQUET_EXPORT Scanner {
 };
 
 template <typename DType>
-class PARQUET_TEMPLATE_CLASS_EXPORT TypedScanner : public Scanner {
+class PARQUET_EXPORT TypedScanner : public Scanner {
  public:
   typedef typename DType::c_type T;