--- a/cpp/src/parquet/util/visibility.h
+++ b/cpp/src/parquet/util/visibility.h
@@ -54,4 +54,14 @@
 #endif
 #endif  // Non-Windows
 
+// This is a complicated topic, some reading on it:
+// http://www.codesynthesis.com/~boris/blog/2010/01/18/dll-export-cxx-templates/
+#if defined(_MSC_VER) || defined(__clang__)
+#define PARQUET_TEMPLATE_CLASS_EXPORT
+#define PARQUET_TEMPLATE_EXPORT PARQUET_EXPORT
+#else
+#define PARQUET_TEMPLATE_CLASS_EXPORT PARQUET_EXPORT
+#define PARQUET_TEMPLATE_EXPORT
+#endif
+
 #endif  // PARQUET_UTIL_VISIBILITY_H