This patch supersedes https://github.com/apache/arrow/pull/3752

I took the liberty of consolidating the comparator code with the statistics code since the two things are effectively inseparable. I also renamed the statistics classes for clarity, since "Statistics" is clearer than "RowGroupStatistics" -- the "scope" of the statistics need not be limited to a row group. 

I apologize for the size of the diff; it is largely the result of moving code around and shuffling code from header files into parquet/statistics.cc