@wesm 2019-05-01
I wanted to get this up for feedback in case I did something deeply offensive -- I will add doxygen comments to the header files before merging this.

I also need to run the benchmarks to see if there is impact (this should be faster, but I'm not sure how much the benchmarks exercise the stats on the write path)