@wesm 2019-05-01
I'll fix the CI issues tomorrow.

Write performance benchmarks

before

```
--------------------------------------------------------------------------------------------------------------
Benchmark                                                                       Time           CPU Iterations
--------------------------------------------------------------------------------------------------------------
BM_WriteInt64Column<Repetition::REQUIRED>/1048576                         6112498 ns    6112346 ns        117   327.207MB/s
BM_WriteInt64Column<Repetition::OPTIONAL>/1048576                         8175286 ns    8175149 ns         86   244.644MB/s
BM_WriteInt64Column<Repetition::REPEATED>/1048576                        11043673 ns   11043294 ns         64   181.105MB/s
BM_WriteInt64Column<Repetition::REQUIRED, Compression::SNAPPY>/1048576    6491432 ns    6491125 ns        107   308.113MB/s
BM_WriteInt64Column<Repetition::OPTIONAL, Compression::SNAPPY>/1048576    8357793 ns    8357599 ns         84   239.303MB/s
BM_WriteInt64Column<Repetition::REPEATED, Compression::SNAPPY>/1048576   11278808 ns   11278491 ns         63   177.329MB/s
BM_WriteInt64Column<Repetition::REQUIRED, Compression::LZ4>/1048576       6614531 ns    6614423 ns        105    302.37MB/s
BM_WriteInt64Column<Repetition::OPTIONAL, Compression::LZ4>/1048576       8529249 ns    8528994 ns         81   234.494MB/s
BM_WriteInt64Column<Repetition::REPEATED, Compression::LZ4>/1048576      11289507 ns   11289159 ns         62   177.161MB/s
BM_WriteInt64Column<Repetition::REQUIRED, Compression::ZSTD>/1048576      6566587 ns    6565922 ns        106   304.603MB/s
BM_WriteInt64Column<Repetition::OPTIONAL, Compression::ZSTD>/1048576      8647931 ns    8647881 ns         81   231.271MB/s
BM_WriteInt64Column<Repetition::REPEATED, Compression::ZSTD>/1048576     11137814 ns   11137516 ns         62   179.573MB/s
```

after

```
--------------------------------------------------------------------------------------------------------------
Benchmark                                                                       Time           CPU Iterations
--------------------------------------------------------------------------------------------------------------
BM_WriteInt64Column<Repetition::REQUIRED>/1048576                         4245946 ns    4245849 ns        166   471.048MB/s
BM_WriteInt64Column<Repetition::OPTIONAL>/1048576                         6416929 ns    6416840 ns        108    311.68MB/s
BM_WriteInt64Column<Repetition::REPEATED>/1048576                         9377082 ns    9376758 ns         75   213.293MB/s
BM_WriteInt64Column<Repetition::REQUIRED, Compression::SNAPPY>/1048576    4551200 ns    4551132 ns        156   439.451MB/s
BM_WriteInt64Column<Repetition::OPTIONAL, Compression::SNAPPY>/1048576    6406567 ns    6406357 ns        109    312.19MB/s
BM_WriteInt64Column<Repetition::REPEATED, Compression::SNAPPY>/1048576    9436338 ns    9436170 ns         74    211.95MB/s
BM_WriteInt64Column<Repetition::REQUIRED, Compression::LZ4>/1048576       4488158 ns    4487987 ns        158   445.634MB/s
BM_WriteInt64Column<Repetition::OPTIONAL, Compression::LZ4>/1048576       6364199 ns    6364079 ns        109   314.264MB/s
BM_WriteInt64Column<Repetition::REPEATED, Compression::LZ4>/1048576       9488606 ns    9488350 ns         75   210.785MB/s
BM_WriteInt64Column<Repetition::REQUIRED, Compression::ZSTD>/1048576      4764453 ns    4764293 ns        149   419.789MB/s
BM_WriteInt64Column<Repetition::OPTIONAL, Compression::ZSTD>/1048576      6514745 ns    6514404 ns        110   307.012MB/s
BM_WriteInt64Column<Repetition::REPEATED, Compression::ZSTD>/1048576      9441827 ns    9441712 ns         76   211.826MB/s
```

So anywhere from 10-40% faster with the vectorized stats