@wesm 2019-05-02
@pitrou @lidavidm I fixed some (valid) clang warnings that aren't caught in our CI for some reason -- I guess we don't build Flight with clang yet