@wesm 2019-05-02
This patch is ready to go I think, I'm just addressing the last CI issues